I changed ProcessAttributes to use intVector instead of doubleVector for pids and ppids. I don't know why I went with doubleVector originally.
I re-enabled the ability to build ITAPS_C plugin itself
(not a copy) against an iMesh implementation. This is to
support a development build as opposed to an installation
of VisIt. I added an example CMake setup for this to
hoth config-site file.




git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@9526 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/viewer/main/ViewerSubject.C
@@ -8917,6 +8917,8 @@ ViewerSubject::ResetInteractorAttributes()
 //   Brad Whitlock, Mon Feb 12 12:11:16 PDT 2007
 //   Made it use ViewerState.
 //
+//   Mark C. Miller, Wed Jan 20 16:41:24 PST 2010
+//   Changed pids, ppids of ProcessAttributes to intVectors.
 // ****************************************************************************
 
 void
@@ -8945,10 +8947,10 @@ ViewerSubject::GetProcessAttributes()
         char myHost[256];
         gethostname(myHost, sizeof(myHost));
 
-        std::vector<double> tmpPids;
+        std::vector<int> tmpPids;
         tmpPids.push_back(pid);
 
-        std::vector<double> tmpPpids;
+        std::vector<int> tmpPpids;
         tmpPpids.push_back(ppid);
 
         std::vector<string> tmpHosts;
